+++ EMS/src/TimeSheet.Modules.EmploymentManagement.Application/Authentication/Renew/Command.cs
@@ -1,0 +1,3 @@
+namespace TimeSheet.Modules.EmploymentManagement.Application.Authentication.Renew;
+
+public sealed record Command();