+++ EMS/src/TimeSheet.Modules.EmploymentManagement.Application/Authentication/Renew/Validator.cs
@@ -1,0 +1,7 @@
+using FluentValidation;
+
+namespace TimeSheet.Modules.EmploymentManagement.Application.Authentication.Renew;
+
+public sealed class Validator : AbstractValidator<Command>
+{
+}